Move code from __main__ to test.py, main.py.
__main__ imports app() from main now.

test.py can be run with python -m redj_rss/text.py
